# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Sarangi DMS API
  version: 1.0.0-rc2
  description: Backend API for the Sarangi Dental Management System.
  contact:
    name: Sarangi DMS Support
    email: support@example.com
servers:
  - url: https://api.dev.example.com
    description: Dev
  - url: https://api.staging.example.com
    description: Staging
  - url: https://api.example.com
    description: Production
security:
  - bearerAuth: []
...
paths:
...
  /xrays/presign:
    post:
      tags: [Xrays]
      summary: Get presigned upload URL
      description: >
        Validates visit and patient, then returns a presigned URL to upload an X-ray file.
      operationId: presignXrayUpload
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/XrayPresignRequest'
      responses:
        '201':
          description: Upload URL
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/XrayPresignResponse'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Visit or patient not found / deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '503':
          description: Presign failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /xrays/{xrayId}/url:
    get:
      tags: [Xrays]
      summary: Get signed download URL for X-ray
      description: >
        Returns a presigned download URL for an X-ray image, for original or thumbnail size.
      operationId: getXrayUrl
      parameters:
        - in: path
          name: xrayId
          required: true
          schema:
            type: string
        - in: query
          name: size
          required: false
          schema:
            type: string
            enum: [thumb, original]
            default: original
      responses:
        '200':
          description: Signed URL
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SignedUrlResponse'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: X-ray/visit/patient not found/deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '503':
          description: URL generation failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
        message:
          type: string
          description: Human-readable error message.
        fieldErrors:
          type: object
          additionalProperties:
            type: array
            items:
              type: string
          description: Optional field-level validation errors.
        traceId:
          type: string
...
    XrayPresignRequest:
      type: object
      required: [visitId, contentType, size]
      properties:
        visitId:
          type: string
        contentType:
          type: string
          enum: [image/jpeg, image/png]
        size:
          type: integer
          minimum: 1024
          maximum: 10485760
    XrayPresignResponse:
      type: object
      required: [xrayId, key, uploadUrl, Headers, expiresInSeconds]
      properties:
        xrayId:
          type: string
        key:
          type: string
        uploadUrl:
          type: string
          format: uri
        Headers:
          type: object
          additionalProperties:
            type: string
        expiresInSeconds:
          type: integer
...
    SignedUrlResponse:
      type: object
      required: [url, variant]
      properties:
        url:
          type: string
          format: uri
        variant:
          type: string
          enum: [thumb, original]
...